#1: Added Makefile, Changed style, Organized components into components folder, css files into css folder
