+++ react-basics/Makefile
@@ -1,0 +1,2 @@
+Dev:
+	npm run dev
