--- react-basics/src/App.jsx
@@ -1,20 +1,0 @@
-function App() {
-  return (
-    <>
-      <main>
-        <h1>Fun facts about React</h1>
-        <ol>
-          <li>Was first released in 2013</li>
-          <li>Was originally created by Jordan Walke</li>
-          <li>Has well over 200K stars on GitHub</li>
-        </ol>
-      </main>
-
-      <footer>
-        <small>@ 20xx last_name development. All rights reserved.</small>
-      </footer>
-    </>
-  );
-}
-
-export default App;
