+++ react-basics/src/components/App.jsx
@@ -1,0 +1,13 @@
+import Header from './Header.jsx'
+import Body from './Body.jsx'
+
+function App() {
+  return (
+    <>
+        <Header/>
+        <Body/>
+    </>
+  );
+}
+
+export default App;
