+++ react-basics/src/components/Body.jsx
@@ -1,0 +1,18 @@
+function Body() {
+    return (
+        <main>
+            <div className="facts">
+                <h1 className="title">Fun facts about React</h1>
+                <ul className="factsList">
+                    <li>Was first released in 2013</li>
+                    <li>Was originally created by Jordan Walke</li>
+                    <li>Has well over 200K stars on GitHub</li>
+                    <li>Is maintained by Meta (formerly Facebook)</li>
+                    <li>Powers thousands of enterprise apps, including mobile apps</li>
+                </ul>
+            </div>
+        </main> 
+    )
+}
+
+export default Body;
