+++ react-basics/src/components/Footer.jsx
@@ -1,0 +1,9 @@
+function Footer() {
+    return (
+        <footer>
+            <small>@ 20xx last_name development. All rights reserved.</small>
+        </footer>
+    )
+}
+
+export default Footer;
