+++ react-basics/src/components/Header.jsx
@@ -1,0 +1,14 @@
+function Header() {
+    return (
+        <header className="brand">
+            <div className="brand-logo">
+                <img src="src/assets/react.svg" alt="React Logo" />
+            </div>
+            <div className="brand-name">
+                <h1>ReactFacts</h1>
+            </div>
+        </header>
+    );
+}
+
+export default Header;
