+++ react-basics/src/components/Main.jsx
@@ -1,0 +1,11 @@
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import '../css/index.css'
+import App from './App.jsx'
+
+const root = createRoot(document.getElementById('root'))
+root.render(
+    <StrictMode>
+        <App/>
+    </StrictMode>
+)
