--- react-basics/src/index.css
@@ -1,72 +1,0 @@
-/* General Styles */
-body {
-    padding: 20px;
-    font-family: 'Poppins', Arial, sans-serif;
-    background-color: #f8f9fa; /* Light background for a clean look */
-    color: #212529; /* Dark text for readability */
-    text-align: center;
-}
-
-h1 {
-    margin-top: 30px;
-    font-size: 2rem;
-    font-weight: bold;
-    color: #333;
-}
-
-/* Header */
-.header {
-    display: flex;
-    justify-content: space-between; /* Pushes logo left and nav right */
-    align-items: center;
-    padding: 15px 40px;
-    background-color: #007bff; /* Bootstrap primary blue */
-    color: white;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.logo img {
-    height: 50px; /* Slightly larger logo */
-}
-
-.crumbs ol {
-    list-style: none;
-    display: flex;
-    gap: 25px;
-    padding: 0;
-    margin: 0;
-}
-
-.crumb {
-    cursor: pointer;
-    font-weight: 600;
-    color: white;
-    text-decoration: none;
-    transition: color 0.3s ease;
-}
-
-.crumb:hover {
-    color: #ffdd57; /* Light yellow on hover */
-}
-
-/* Main Content */
-main {
-    padding: 30px;
-    background: white;
-    border-radius: 12px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-    max-width: 600px;
-    margin: 40px auto;
-}
-
-/* Footer */
-footer {
-    margin-top: 30px;
-    padding: 10px;
-    background-color: #343a40;
-    color: white;
-    text-align: center;
-    border-radius: 8px;
-    font-size: 14px;
-}
